LPS-70693 Rename for split package issues.
