+++ adaptive-media-blogs-editor-configuration-static/src/main/java/com/liferay/adaptive/media/blogs/editor/configuration/internal/stat1c/StaticAdaptiveMediaBlogsEditorConfigContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.blogs.editor.configuration.internal;
+package com.liferay.adaptive.media.blogs.editor.configuration.internal.stat1c;
 
 import com.liferay.adaptive.media.image.item.selector.AdaptiveMediaImageURLItemSelectorReturnType;
 import com.liferay.blogs.item.selector.criterion.BlogsItemSelectorCriterion;
